+++ src/lib.rs
@@ -1,19 +1,21 @@
 pub mod consts;
 pub mod editor;
+pub mod gol_utils;
 
+use crate::gol_utils::{build_random, step};
 use std::collections::HashSet;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 
 use consts::*;
 
 use editor::GUIEvent;
+use gol_utils::build_ir;
 use nih_plug::prelude::*;
 use nih_plug_vizia::ViziaState;
 use rand::rngs::SmallRng;
-use rand::{Rng, SeedableRng};
+use rand::SeedableRng;
 use realfft::num_complex::Complex;
 use realfft::{ComplexToReal, RealFftPlanner, RealToComplex};
-use rtrb::*;
 
 struct Automata {
     params: Arc<AutomataParams>,
@@ -24,13 +26,19 @@ struct Automata {
     stft: util::StftHelper,
 
     comp_buff: Vec<Complex<f32>>,
+    real_buff: Vec<f32>,
 
-    ir_cons: Option<Consumer<Complex<f32>>>,
     current_ir: Vec<Complex<f32>>,
+    current_board: HashSet<(i32, i32)>,
+    dying_buff: Vec<(i32, i32)>,
+    born_buff: Vec<(i32, i32)>,
 }
 
 #[derive(Params)]
 struct AutomataParams {
+    #[id = "running"]
+    running: BoolParam,
+
     #[persist = "editor-state"]
     editor_state: Arc<ViziaState>,
 }
@@ -41,13 +49,26 @@ impl Default for Automata {
         let fft = planner.plan_fft_forward(FFT_WINDOW_SIZE);
         let ifft = planner.plan_fft_inverse(FFT_WINDOW_SIZE);
 
-        let comp_buff = ifft.make_input_vec();
+        let mut comp_buff = ifft.make_input_vec();
+        let mut real_buff = fft.make_input_vec();
+
+        let mut current_board: HashSet<(i32, i32)> =
+            HashSet::with_capacity(FILTER_WINDOW_SIZE * FILTER_WINDOW_SIZE);
+        let mut rng = SmallRng::seed_from_u64(SEED);
+        let born_buff = Vec::with_capacity(FILTER_WINDOW_SIZE * FILTER_WINDOW_SIZE);
+        let dying_buff = Vec::with_capacity(FILTER_WINDOW_SIZE * FILTER_WINDOW_SIZE);
+
+        build_random(&mut current_board, &mut rng);
+        build_ir(&mut current_board, &mut real_buff);
+
+        fft.process_with_scratch(&mut real_buff, &mut comp_buff, &mut [])
+            .unwrap();
 
         Self {
             params: Arc::new(AutomataParams::default()),
 
             current_ir: comp_buff.clone(),
-            ir_cons: None,
+            current_board: HashSet::with_capacity(FILTER_WINDOW_SIZE * FILTER_WINDOW_SIZE),
 
             fft,
             ifft,
@@ -55,6 +76,9 @@ impl Default for Automata {
             stft: util::StftHelper::new(2, WINDOW_SIZE, FFT_WINDOW_SIZE - WINDOW_SIZE),
 
             comp_buff,
+            real_buff,
+            born_buff,
+            dying_buff,
         }
     }
 }
@@ -63,6 +87,7 @@ impl Default for AutomataParams {
     fn default() -> Self {
         Self {
             editor_state: editor::default_state(),
+            running: BoolParam::new("running", false),
         }
     }
 }
@@ -104,170 +129,6 @@ impl Plugin for Automata {
     // tasks.
 
     type BackgroundTask = GUIEvent;
-
-    // TODO move this shit to the audio thread
-    fn task_executor(&mut self) -> TaskExecutor<Self> {
-        let (ir_prod, ir_cons) = rtrb::RingBuffer::<Complex<f32>>::new(FILTER_WINDOW_SIZE * 3);
-        self.ir_cons = Some(ir_cons);
-
-        let celled = Arc::new(Mutex::new(ir_prod));
-
-        Box::new(move |_| {
-            panic!();
-
-            let mut prod = match celled.lock() {
-                Ok(x) => x,
-                Err(_) => todo!(),
-            };
-
-            // initialize stuff
-            let mut current_board: HashSet<(i32, i32)> =
-                HashSet::with_capacity(FILTER_WINDOW_SIZE * FILTER_WINDOW_SIZE);
-            let mut rng = SmallRng::seed_from_u64(SEED);
-            let mut planner = RealFftPlanner::new();
-            let real_to_complex = planner.plan_fft_forward(FFT_WINDOW_SIZE);
-
-            let mut comp_buff = real_to_complex.make_output_vec();
-            let mut real_buff = real_to_complex.make_input_vec();
-
-            // make some closures
-            let mut build_random = |board: &mut HashSet<(i32, i32)>| {
-                for i in 0..FILTER_WINDOW_SIZE {
-                    for j in 0..FILTER_WINDOW_SIZE {
-                        if rng.gen() {
-                            board.insert((i as i32, j as i32));
-                        }
-                    }
-                }
-            };
-            let find_neighbors = |pos: &(i32, i32)| -> Vec<(i32, i32)> {
-                let mut neighbors: Vec<(i32, i32)> = Vec::new();
-                for x in -1..2 {
-                    for y in -1..2 {
-                        if x != 0 || y != 0 {
-                            neighbors.push((pos.0 + x, pos.1 + y));
-                        }
-                    }
-                }
-                neighbors
-            };
-            let build_ir = |board: &HashSet<(i32, i32)>,
-                            comp_buff: &mut Vec<Complex<f32>>,
-                            real_buff: &mut Vec<f32>| {
-                let mut ir: Vec<f32> = vec![0.0; FILTER_WINDOW_SIZE];
-                for i in 0..FILTER_WINDOW_SIZE {
-                    ir[i] = {
-                        let mut out = 0.0;
-                        for j in 0..FILTER_WINDOW_SIZE {
-                            let b_ij = match board.contains(&(i as i32, j as i32)) {
-                                true => 1.0,
-                                false => -1.0,
-                            };
-                            let b_ji = match board.contains(&(j as i32, i as i32)) {
-                                true => 1.0,
-                                false => -1.0,
-                            };
-
-                            out += b_ij + b_ji;
-                        }
-
-                        out /= FILTER_WINDOW_SIZE as f32;
-                        out
-                    }
-                }
-
-                real_buff[0..FILTER_WINDOW_SIZE].copy_from_slice(&ir[0..FILTER_WINDOW_SIZE]);
-
-                // TODO might want to think about moving the fft to the audio thread
-                // and just sending the real ir over
-                real_to_complex
-                    .process_with_scratch(real_buff, comp_buff, &mut [])
-                    .unwrap();
-            };
-
-            // initialize a random board
-            build_random(&mut current_board);
-
-            build_ir(&mut current_board, &mut comp_buff, &mut real_buff);
-            match prod.write_chunk(FILTER_WINDOW_SIZE) {
-                Ok(mut chunk) => {
-                    let slices = chunk.as_mut_slices();
-
-                    let first_len = slices.0.len();
-                    let second_len = slices.0.len();
-
-                    slices.0[0..first_len].copy_from_slice(&comp_buff[0..first_len]);
-                    slices.1[0..second_len]
-                        .copy_from_slice(&comp_buff[first_len..first_len + second_len]);
-
-                    chunk.commit_all();
-                }
-                Err(_) => {
-                    todo!();
-                }
-            }
-            comp_buff.clear();
-
-            let mut dying: Vec<(i32, i32)> = Vec::new();
-            let mut born: Vec<(i32, i32)> = Vec::new();
-
-            loop {
-                for cell in current_board.iter() {
-                    let neighbors = find_neighbors(&cell);
-                    let mut living_neighbors: u8 = 0;
-                    for neighbor in neighbors.iter() {
-                        if current_board.contains(neighbor) {
-                            living_neighbors += 1;
-                        } else if !born.contains(neighbor) {
-                            let neighbors_neighbors = find_neighbors(neighbor);
-                            let mut neighbors_living_neighbors: u8 = 0;
-                            for neighbor_neighbor in neighbors_neighbors.iter() {
-                                if current_board.contains(neighbor_neighbor) {
-                                    neighbors_living_neighbors += 1;
-                                }
-                            }
-                            if neighbors_living_neighbors == 3 && !born.contains(neighbor) {
-                                born.push(*neighbor);
-                            }
-                        }
-                    }
-                    if living_neighbors > 3 || living_neighbors < 2 && !dying.contains(cell) {
-                        dying.push(*cell);
-                    }
-                }
-                for cell in dying.iter() {
-                    current_board.remove(cell);
-                }
-                for cell in born.iter() {
-                    current_board.insert(*cell);
-                }
-                dying.clear();
-                born.clear();
-
-                // build impluse response and send it to audio thread
-                build_ir(&mut current_board, &mut comp_buff, &mut real_buff);
-                match prod.write_chunk(FILTER_WINDOW_SIZE * 5) {
-                    Ok(mut chunk) => {
-                        let slices = chunk.as_mut_slices();
-
-                        let first_len = slices.0.len();
-                        let second_len = slices.0.len();
-
-                        slices.0[0..first_len].copy_from_slice(&comp_buff[0..first_len]);
-                        slices.1[0..second_len]
-                            .copy_from_slice(&comp_buff[first_len..first_len + second_len]);
-
-                        chunk.commit_all();
-                    }
-                    Err(_) => {
-                        todo!();
-                    }
-                }
-
-                comp_buff.clear();
-            }
-        })
-    }
 
     fn params(&self) -> Arc<dyn Params> {
         self.params.clone()
@@ -314,24 +175,20 @@ impl Plugin for Automata {
     ) -> ProcessStatus {
         nih_log!("doing a process block");
 
-        // TODO figure out how to handle panic here
-        match self
-            .ir_cons
-            .as_mut()
-            .unwrap()
-            .read_chunk(FILTER_WINDOW_SIZE)
-        {
-            Ok(ir) => {
-                let slices = ir.as_slices();
-                self.current_ir[0..slices.0.len()].copy_from_slice(slices.0);
-                self.current_ir[slices.0.len()..slices.0.len() + slices.1.len()]
-                    .copy_from_slice(slices.1);
-                ir.commit_all()
-            }
-            Err(_e) => {
-                // panic!()
-            }
-        };
+        if self.params.running.value() {
+            step(
+                &mut self.current_board,
+                &mut self.born_buff,
+                &mut self.dying_buff,
+                &mut self.real_buff,
+            );
+
+            self.fft
+                .process_with_scratch(&mut self.real_buff, &mut self.comp_buff, &mut [])
+                .unwrap();
+
+            self.current_ir.copy_from_slice(&self.comp_buff);
+        }
 
         self.stft
             .process_overlap_add(buffer, 1, |_channel, real_buff| {
